+++ arangod/Aql/ExecutionNode/IResearchViewNode.h
@@ -28,8 +28,7 @@
 #include "Aql/ExecutionNode/DataAccessingNode.h"
 #include "Aql/ExecutionNode/ExecutionNode.h"
 #include "Aql/ExecutionNodeId.h"
-#include "Aql/LateMaterializedOptimizerRulesCommon.h"
-#include "Aql/types.h"
+#include "Aql/Optimizer/Rule/OptimizerRulesLateMaterializedCommon.h"
 #include "Aql/types.h"
 #include "Containers/FlatHashMap.h"
 #include "Containers/FlatHashSet.h"
