+++ arangod/Aql/Optimizer/Rule/OptimizerRuleBatchMaterializeDocuments.cpp
