+++ arangod/Aql/Optimizer/Rule/OptimizerRulePropagateConstantAttributes.cpp
@@ -29,11 +29,11 @@
 #include "Aql/ExecutionNode/SubqueryNode.h"
 #include "Aql/Expression.h"
 #include "Aql/Optimizer.h"
+#include "Aql/OptimizerRules.h"
+#include "Aql/OptimizerUtils.h"
 #include "Cluster/ServerState.h"
 #include "Containers/ImmutableMap.h"
 #include "Logger/LogMacros.h"
-#include "OptimizerRules.h"
-#include "OptimizerUtils.h"
 #include "VocBase/LogicalCollection.h"
 
 #include <boost/container_hash/hash.hpp>
