+++ arangod/Aql/Optimizer/Rule/OptimizerRulesJoin.cpp
@@ -38,6 +38,7 @@
 #include "Aql/Expression.h"
 #include "Aql/IndexStreamIterator.h"
 #include "Aql/Optimizer.h"
+#include "Aql/OptimizerRules.h"
 #include "Aql/OptimizerUtils.h"
 #include "Aql/Projections.h"
 #include "Aql/Query.h"
@@ -55,7 +56,6 @@
 #include "Graph/TraverserOptions.h"
 #include "Indexes/Index.h"
 #include "Logger/LogMacros.h"
-#include "OptimizerRules.h"
 #include "Utils/CollectionNameResolver.h"
 #include "VocBase/LogicalCollection.h"
 #include "VocBase/Methods/Collections.h"
